Fix type warnings (size_t - int)

+++ week4_divide_and_conquer/3_improving_quicksort/sorting.cpp
@@ -77,11 +77,11 @@ int main() {
   int n;
   std::cin >> n;
   vector<int> a(n);
-  for (size_t i = 0; i < a.size(); ++i) {
+  for (int i = 0; i < ((int) a.size()); ++i) {
     std::cin >> a[i];
   }
-  randomized_quick_sort3(a, 0, a.size() - 1);
-  for (size_t i = 0; i < a.size(); ++i) {
+  randomized_quick_sort3(a, 0, (int) a.size() - 1);
+  for (int i = 0; i < ((int) a.size()); ++i) {
     std::cout << a[i] << ' ';
   }
 }
